+++ src/roswire/app/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from .app import App
+from .description import AppDescription
+from .instance import AppInstance