+++ src/roswire/node/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from .manager import NodeManager
+from .node import Node