+++ src/roswire/service/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from .manager import ServiceManager
+from .service import Service